+++ src/yulplus.js
@@ -115,8 +115,6 @@ function id(x) { return x[0]; }
 
     // if mstore(0, x, x, x) args, than process this..
     if (d[0].value === 'mstore' && d[2][3].length > 1) {
-      console.log('special mstore', d);
-
       // values after pos, x, [....]
       const secondaryValues = d[2][3].slice(1);
 
@@ -454,7 +452,6 @@ var grammar = {
               value: '',
               text: '',
               _includeMarker: 'mslice',
-              usesSafeMath: true,
               toString: () => '',
             });
           }
@@ -727,7 +724,7 @@ var grammar = {
             dataMap[name + '.size'] = {
               method: `
         function ${name + '.size'}(pos) -> _offset {
-          _offset := safeSub(${name + '.offset'}(pos), pos)
+          _offset := sub(${name + '.offset'}(pos), pos)
         }
         `,
               required: [name + '.offset'],