+++ src/yulplus.js
@@ -32,7 +32,7 @@ function id(x) { return x[0]; }
     boolean: ["true", "false"],
     bracket: ["{", "}", "(", ")", '[', ']'],
     ConstIdentifier: /(?:const)(?:\s)/,
-    keyword: ['code', 'let', "for", "function", "enum", "mstruct",
+    keyword: ['code ', 'let', "for", "function", "enum", "mstruct",
       "if", "else", "break", "continue", "default", "switch", "case"],
     Identifier: /[\w.]+/,
   });